backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
apscheduler
tesseract-ocr
google-auth
